Rename the driver-file to driver-filestream.
