+++ Dockerfile
@@ -31,7 +31,7 @@ COPY --chown=root:root sbk-api ${APP_HOME}/sbk-api
 COPY --chown=root:root driver-hdfs ${APP_HOME}/driver-hdfs
 COPY --chown=root:root driver-bookkeeper ${APP_HOME}/driver-bookkeeper
 COPY --chown=root:root driver-concurrentq ${APP_HOME}/driver-concurrentq
-COPY --chown=root:root driver-file ${APP_HOME}/driver-file
+COPY --chown=root:root driver-filestream ${APP_HOME}/driver-filestream
 COPY --chown=root:root driver-asyncfile ${APP_HOME}/driver-asyncfile
 COPY --chown=root:root driver-filechannel ${APP_HOME}/driver-filechannel
 COPY --chown=root:root driver-kafka ${APP_HOME}/driver-kafka
