+++ driver-filestream/build.gradle
@@ -3,8 +3,6 @@ plugins {
 }
 
 
-sourceCompatibility = 1.8
-
 repositories {
     mavenCentral()
 }
