+++ settings.gradle
@@ -15,8 +15,8 @@ include 'sbk-api'
 include 'driver-pulsar'
 include 'driver-kafka'
 include 'driver-pravega'
+include 'driver-filestream'
 include 'driver-asyncfile'
-include 'driver-file'
 include 'driver-filechannel'
 include 'driver-concurrentq'
 include 'driver-bookkeeper'
